chore: Address feedback. Move `generateUUID` to a utils file
