+++ src/main.ts
@@ -63,6 +63,7 @@ import {
 import { SDKEventName } from "./behavioural-events/sdk-events";
 import { autoParseUTMParams } from "./helpers/utm-params";
 import { defaultFlagsConfig, type FlagsConfig } from "./entities/flags-config";
+import { generateUUID } from "./helpers/uuid-helper";
 
 export { ProductType } from "./entities/offerings";
 export type {
@@ -771,34 +772,6 @@ export class Purchases {
    * @public
    */
   public static generateRevenueCatAnonymousAppUserId(): string {
-    const generateUUID = (): string => {
-      if (crypto && crypto.randomUUID) {
-        return crypto.randomUUID();
-      }
-
-      const bytes = new Uint8Array(16);
-      if (crypto && crypto.getRandomValues) {
-        crypto.getRandomValues(bytes);
-      } else {
-        for (let i = 0; i < 16; i++) {
-          bytes[i] = Math.floor(Math.random() * 256);
-        }
-      }
-
-      // Set version (4) and variant bits
-      bytes[6] = (bytes[6] & 0x0f) | 0x40; // version 4
-      bytes[8] = (bytes[8] & 0x3f) | 0x80; // variant
-
-      const hex = Array.from(bytes).map((b) => b.toString(16).padStart(2, "0"));
-      return [
-        hex.slice(0, 4).join(""),
-        hex.slice(4, 6).join(""),
-        hex.slice(6, 8).join(""),
-        hex.slice(8, 10).join(""),
-        hex.slice(10, 16).join(""),
-      ].join("-");
-    };
-
     return `${ANONYMOUS_PREFIX}${generateUUID().replace(/-/g, "")}`;
   }
